SIM-17409 simQt::AsyncCategoryCounter Now Resets Correctly
**JIRA Issue:** SIM-17409

**Description:** The lack of a reset allowed old filter settings to be applied to a new scenario.

**Testing Performed:** Before the change I could consistently get incorrect category counts, now the counts are always correct.

+++ SDK/simQt/CategoryTreeModel.cpp
@@ -1242,6 +1242,10 @@ void CategoryTreeModel::setSettings(Settings* settings, const QString& settingsK
 
 void CategoryTreeModel::clearTree_()
 {
+  // only reset if necessary
+  if ((categories_.size() == 0) && categoryIntToItem_.empty() && filter_->isEmpty())
+    return;
+
   beginResetModel();
   categories_.deleteAll();
   categoryIntToItem_.clear();
